feat(home): cross-dissolve home states + animate the re-plan, smoother first load
- HomeScreen now crossfades between loading / idle / plan layouts (gated on `initialized`)
  instead of hard-cutting: cold start fades straight into the resolved layout (no empty→
  thread flash / card pop), and idle↔plan switches dissolve. Extracted the idle and plan
  branches into HomeIdleContent / HomePlanContent (the latter + StickyAlarm moved to a new
  HomeContent.kt to stay under the file-length cap).
- AiThinkingThread animates the answer in/out (AnimatedVisibility: alpha-in, shrink-out) so
  a re-plan collapses the old answer smoothly into "Thinking…" instead of jumping; fadeIn is
  alpha-only so it never fights the streaming reveal, and the last answer is held for the exit.
